feat: added types to json rules
